Connected to mongodb atlas

+++ index.js
@@ -3,13 +3,20 @@ const bodyParser = require("body-parser");
 const getDate = require(__dirname + "/date.js");
 const app = express();
 const mongoose = require("mongoose");
+const _ = require("lodash");
+require('dotenv').config();
 const port = 3000;
 
 app.use(bodyParser.urlencoded({extended:true}));
 app.set('view engine', 'ejs');
 app.use(express.static("public"));
 
-mongoose.connect("mongodb://localhost:27017/todolistDB", {useNewUrlParser: true, useUnifiedTopology: true});
+const year = getDate.getYear();
+
+const dbUser = process.env.DB_USER;
+const dbPass = process.env.DB_PASS
+
+mongoose.connect(`mongodb+srv://${dbUser}:${dbPass}@cluster0-qnmpu.mongodb.net/test?retryWrites=true&w=majority/todolistDB`, {useNewUrlParser: true, useUnifiedTopology: true, useFindAndModify: false});
 
 const itemSchema = {
     name: String
@@ -34,8 +41,6 @@ const listSchema = {
 
 const List = mongoose.model("list", listSchema);
 
-year = getDate.getYear();
-
 app.get("/", (req, res) => {
 
     Item.find({}, (err, result) => {
@@ -49,8 +54,8 @@ app.get("/", (req, res) => {
                 res.redirect("/");
             });
         } else{
-                day = getDate.getDate();
-                res.render('list', {listType: day, year: year, newListItems: result});    
+                
+                res.render('list', {listType: "Today", year: year,  newListItems: result});    
         }
     });
 
@@ -59,48 +64,67 @@ app.get("/", (req, res) => {
 
 app.post("/", (req, res) => {
     const newItem = req.body.newItem;
-    console.log(newItem);
+    const btnListname = req.body.list;
+
     const item = new Item({
         name: newItem
     });
+
+    if(btnListname === "Today"){
     item.save();
     res.redirect("/");
+    } else {
+        List.findOne({name: btnListname}, (err, newFoundList) => {
+            newFoundList.items.push(item);
+            newFoundList.save();
+            res.redirect("/" + btnListname);
+        });
+    }
 });
 
 app.get("/:listType", (req,res) => {
-    const listName = req.params.listType;
-
-    List.find({listName}, (err, results) => {
-        results.forEach(result => {
-        const testName = result.name;
-        
-        if(testName != listName){
-            const list = new List({
-                name: listName,
-                items: defaultItems
-            });
-        
-            list.save();
-        } else {
-            console.log(testName)
+    const listName = _.capitalize(req.params.listType);
+    
+    List.findOne({name: listName}, (err, listFound) => {
+        if(!err){
+            if(!listFound){
+                const list = new List({
+                    name: listName,
+                    items: defaultItems
+                });
+                list.save();
+                res.redirect("/" + listName);
+            } else {
+                //show an existing list  
+                res.render('list', {listType: listName, year: year, newListItems: listFound.items});             
+            }
         }
     });
-    });
 
-        
+    
 });
 
 
 app.post("/delete", (req, res) => {
     const id = req.body.checkbox;
+    const lname = req.body.hiddenListName;
 
-    Item.findByIdAndRemove(id, {useFindAndModify: false}, (err) => {
-        if(err) {
-            console.log(err);            
-        } else {
-            res.redirect("/");
-        }
-    });
+    if(lname === "Today"){
+        Item.findByIdAndRemove(id, {useFindAndModify: false}, (err) => {
+            if(err) {
+                console.log(err);            
+            } else {
+                res.redirect("/");
+            }
+        });
+    } else {
+        List.findOneAndUpdate({name: lname}, {$pull: {items: {_id: id}}}, (err, delFoundList) => {
+            if(!err){
+                res.redirect("/" + lname);
+            }
+        });
+    }
+    
 });
 
 app.get("/about", (req, res) => {
